refactor(ui): make the DHCP table work for all nodes
Update the DHCP table to handle all node types.

Fixes: canonical-web-and-design/app-tribe#568.
