+++ ui/src/app/base/components/DHCPTable/EditDHCP/EditDHCP.tsx
