+++ ui/src/app/base/components/DHCPTable/EditDHCP/index.ts
