+++ ui/src/app/base/components/DHCPTable/_index.scss
@@ -1,4 +1,4 @@
-@mixin MachineDHCPTable {
+@mixin DHCPTable {
   .dhcp-snippets-table {
     th,
     td {
