+++ ui/src/app/base/components/DHCPTable/index.ts
